# Optimize Status#permitted_for 24x