@alpaca-tc 2017-05-15
Update a pull request. 😄 👍 

- Add spec
- Use `IN` instead of `NOT IN` to optimize query
- Convert symbols to enum integer with Attribute API